Moved sky renderer into renderers folder
