+++ packages/rewild-renderer/lib/core/Atmosphere.ts
@@ -1,5 +1,5 @@
 import { Renderer } from '..';
-import { AtmosphereMaterial } from '../materials/sky/AtmosphereMaterial';
+import { AtmosphereMaterial } from '../renderers/sky/AtmosphereMaterial';
 import { Camera } from './Camera';
 
 export class Atmosphere {
